requirements.txt
atlassian-python-api
jsonschema
pyaml
python-dateutil
python-gitlab
python-jenkins
requests
urllib3
